# Fix StreamReader EOF handling and improve perf