@danmoseley 2022-07-25
@GrabYourPitchforks do you expect to have time to continue this PR? I see that @62552 by @trapov. Is much work remaining?

@stephentoub how important is it that we get these PR's into .NET 7?