@stephentoub 2022-07-25
It'd be valuable to get this PR for .NET 7, but we wouldn't block the release on it :smile: