@Jozkee 2022-08-30
CI issues are related:

```
...
  Discovering: System.IO.Tests (method display = ClassAndMethod, method display options = None)
  Discovered:  System.IO.Tests (found 719 of 727 test cases)
  Starting:    System.IO.Tests (parallel test collections = on, max threads = 4)
Process terminated. Assertion failed.
   at System.IO.StreamReader.ReadAsyncInternal(Memory`1 buffer, CancellationToken cancellationToken) in /_/src/libraries/System.Private.CoreLib/src/System/IO/StreamReader.cs:line 1215
   at System.Runtime.CompilerServices.AsyncMethodBuilderCore.Start[TStateMachine](TStateMachine& stateMachine) in /_/src/libraries/System.Private.CoreLib/src/System/Runtime/CompilerServices/AsyncMethodBuilderCore.cs:line 38
   at System.IO.StreamReader.ReadAsyncInternal(Memory`1 buffer, CancellationToken cancellationToken)
   at System.IO.StreamReader.ReadAsync(Memory`1 buffer, CancellationToken cancellationToken) in /_/src/libraries/System.Private.CoreLib/src/System/IO/StreamReader.cs:line 1080
   at System.IO.Tests.StreamReaderTests.ReadAsync_LongStreamIntoShortBuffer_PerformsFinalFlushCorrectly() in /_/src/libraries/System.IO/tests/StreamReader/StreamReaderTests.cs:line 819
   at System.Runtime.CompilerServices.AsyncMethodBuilderCore.Start[TStateMachine](TStateMachine& stateMachine) in /_/src/libraries/System.Private.CoreLib/src/System/Runtime/CompilerServices/AsyncMethodBuilderCore.cs:line 38
   at System.IO.Tests.StreamReaderTests.ReadAsync_LongStreamIntoShortBuffer_PerformsFinalFlushCorrectly()
   at System.RuntimeMethodHandle.InvokeMethod(Object target, Void** arguments, Signature sig, Boolean isConstructor)
...